[server] Refactor OpenAI server into modular architecture with reasoning support
Split the monolithic openai_server.py (1076 lines) into focused modules
following the mini-sglang/vLLM pattern:

  atom/entrypoints/openai/
  ├── __init__.py           - Package re-export
  ├── protocol.py           - Pydantic request/response models
  ├── reasoning.py          - Thinking content separation (reasoning_content)
  ├── serving_chat.py       - Chat completion handler
  ├── serving_completion.py - Text completion handler
  └── api_server.py         - FastAPI app, routes, engine interface, main()

Key improvements:
- Reasoning content separation: <think>...</think> blocks are now returned
  in a `reasoning_content` field alongside `content`, following the
  SGLang/vLLM pattern (instead of stripping entirely)
- Streaming reasoning: thinking tokens stream as reasoning_content deltas,
  answer tokens as content deltas
- Tool call token stripping: <|tool_calls_section_*|> tokens filtered
- OpenAI-format error responses via exception handlers
- Proper ModelCard/ModelList Pydantic models for /v1/models
- Compatibility fields: presence_penalty, frequency_penalty, n accepted
- Backward compatible: python -m atom.entrypoints.openai_server still works
