+++ atom/entrypoints/openai/__init__.py
@@ -1,0 +1,8 @@
+# SPDX-License-Identifier: MIT
+# Copyright (C) 2024-2025, Advanced Micro Devices, Inc. All rights reserved.
+
+"""ATOM OpenAI-compatible API server package."""
+
+from .api_server import main
+
+__all__ = ["main"]
